refactor: extract route logic and state
